+++ src/json_number.cpp
@@ -23,7 +23,6 @@
 #include <variant>
 #include <terra/json/json.h>
 #include <terra/json/json_exception.h>
-#include "has_format.h"
 #ifdef TERRA_HAS_FORMAT
 #include <format>
 #endif