+++ src/json_parser.cpp
@@ -31,7 +31,6 @@
 #include <terra/json/json_exception.h>
 #include <terra/json/json_parser.h>
 #include "unicode_constants.h"
-#include "has_format.h"
 #ifdef TERRA_HAS_FORMAT
 #include <format>
 #else